backend/requirements/requirements.txt
litestar
uvicorn
jinja2
piccolo[postgres]
piccolo_admin
pydantic_settings